(long long)

+++ ressched/interfaces/innerkits/ressched_client/src/res_sched_client.cpp
@@ -32,7 +32,7 @@ void ResSchedClient::ReportData(uint32_t resType, int64_t value,
                                 const std::unordered_map<std::string, std::string>& mapPayload)
 {
     RESSCHED_LOGD("ResSchedClient::ReportData receive resType = %{public}u, value = %{public}lld.",
-                  resType, value);
+                  resType, (long long)value);
     if (TryConnect() != ERR_OK) {
         return;
     }
